Add ProveCaller and ActionCaller
The new ProveCaller and the ParseCaller share some code which is
extracted to the new ActionCaller.
