+++ lib/hets/parse_caller.rb
@@ -1,5 +1,5 @@
 module Hets
-  class ParseCaller < Caller
+  class ParseCaller < ActionCaller
     # Server-Instances should be started with the following
     # arguments: `hets +RTS -K1G -RTS -a none`
     MODE_ARGUMENTS = {
@@ -10,15 +10,6 @@ module Hets
     CMD = 'dg'
     METHOD = :get
 
-    attr_accessor :url_catalog
-
-    def initialize(hets_instance, url_catalog = [])
-      self.url_catalog = url_catalog
-      msg = "<#{hets_instance}> not up."
-      raise Hets::InactiveInstanceError, msg unless hets_instance.try(:up?)
-      super(hets_instance)
-    end
-
     def call(iri, with_mode: :default)
       escaped_iri = Rack::Utils.escape_path(iri)
       arguments = [escaped_iri, *MODE_ARGUMENTS[with_mode]]
@@ -26,18 +17,6 @@ module Hets
       perform(api_uri, {}, METHOD)
     rescue UnfollowableResponseError => error
       handle_possible_hets_error(error)
-    end
-
-    def build_query_string
-      query_hash = {}
-      query_hash[:"url-catalog"] = url_catalog.join(',') if url_catalog.present?
-      query_hash
-    end
-
-    def handle_possible_hets_error(error)
-      HetsErrorProcess.new(error).handle
-    rescue Hets::NotAHetsError
-      raise error
     end
   end
 end
